Add UMD build and test HTML page
The rollup configuration has been updated to include a UMD (Universal Module Definition) build for the browser. Also, a new `test.html` file has been added under the `example` directory for testing. The package.json and package-lock.json versions were updated to 0.0.34. The library is now also attached to the window object as `AsyncpayCheckout` for global use. Also update the checkout feature to prevent redundant appending of the public key.